// apps/creator-hub/modules/analytics-assistant/types/AnalyticsAssistantVoteOptions.ts
// NOTE(lucaswang, 2025-03-28): Mirrors https://github.rbx.com/Roblox/creator-feedback/blob/master/services/feedback-gateway-api/src/Models/Enum/AnalyticsAssistantVoteOptions.cs
enum AnalyticsAssistantVoteOptions {
  UpvoteIdentifyNextAction = 'UpvoteIdentifyNextAction',
  UpvoteUnderstanding = 'UpvoteUnderstanding',
  UpvoteSaveTime = 'UpvoteSaveTime',
  UpvoteOther = 'UpvoteOther',
  DownvoteInaccurateData = 'DownvoteInaccurateData',
  DownvoteNotHelpful = 'DownvoteNotHelpful',
  DownvoteNotRelevant = 'DownvoteNotRelevant',
  DownvoteOther = 'DownvoteOther',
  DownvoteInaccurateComments = 'DownvoteInaccurateComments',
  DownvoteMissingImportantComments = 'DownvoteMissingImportantComments',
}

export default AnalyticsAssistantVoteOptions;